+++ static/js/shared/config-help-tips.js
@@ -38,6 +38,7 @@
                 'tipEditUndo',
                 'tipEditDrag',
                 'tipEditRenamePage',
+                'tipEditPageTabDot',
                 'tipEditCopyUrl',
             ],
         },